+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for cabal-scaffold
 
+## Unreleased
+
+- Introduced `project-file` field in config file and deleted `noProject` field.
+
 ## 0.1.0.0
 
 - Initial Release!